+++ packages/github/src/github.ts
@@ -1,27 +1,20 @@
-// Originally pulled from https://github.com/JasonEtco/actions-toolkit/blob/master/src/github.ts
-import {graphql} from '@octokit/graphql'
-
-// we need this type to set up a property on the GitHub object
-// that has token authorization
-// (it is not exported from octokit by default)
-import {graphql as GraphQL} from '@octokit/graphql/dist-types/types'
-
-import Octokit from '@octokit/rest'
 import * as Context from './context'
+import {GitHub, getOctokitOptions} from './utils'
 
-// We need this in order to extend Octokit
-Octokit.prototype = new Octokit()
+// octokit + plugins
+import {OctokitOptions} from '@octokit/core/dist-types/types'
 
 export const context = new Context.Context()
 
-export class GitHub extends Octokit {
-  graphql: GraphQL
-
-  constructor(token: string, opts: Omit<Octokit.Options, 'auth'> = {}) {
-    super({...opts, auth: `token ${token}`})
-
-    this.graphql = graphql.defaults({
-      headers: {authorization: `token ${token}`}
-    })
-  }
+/**
+ * Returns a hydrated octokit ready to use for GitHub Actions
+ *
+ * @param     token    the repo PAT or GITHUB_TOKEN
+ * @param     options  other options to set
+ */
+export function getOctokit(
+  token: string,
+  options?: OctokitOptions
+): InstanceType<typeof GitHub> {
+  return new GitHub(getOctokitOptions(token, options))
 }