+++ packages/tool-cache/__tests__/tool-cache.test.ts
@@ -1,8 +1,9 @@
 import * as fs from 'fs'
-import * as nock from 'nock'
 import * as path from 'path'
 import * as io from '@actions/io'
 import * as exec from '@actions/exec'
+import * as stream from 'stream'
+import nock from 'nock'
 
 const cachePath = path.join(__dirname, 'CACHE')
 const tempPath = path.join(__dirname, 'TEMP')
@@ -14,6 +15,7 @@ process.env['RUNNER_TOOL_CACHE'] = cachePath
 import * as tc from '../src/tool-cache'
 
 const IS_WINDOWS = process.platform === 'win32'
+const IS_MAC = process.platform === 'darwin'
 
 describe('@actions/tool-cache', function() {
   beforeAll(function() {
@@ -24,6 +26,8 @@ describe('@actions/tool-cache', function() {
         username: 'abc',
         password: 'def'
       })
+    setGlobal('TEST_DOWNLOAD_TOOL_RETRY_MIN_SECONDS', 0)
+    setGlobal('TEST_DOWNLOAD_TOOL_RETRY_MAX_SECONDS', 0)
   })
 
   beforeEach(async function() {
@@ -33,9 +37,15 @@ describe('@actions/tool-cache', function() {
     await io.mkdirP(tempPath)
   })
 
+  afterEach(function() {
+    setResponseMessageFactory(undefined)
+  })
+
   afterAll(async function() {
     await io.rmRF(tempPath)
     await io.rmRF(cachePath)
+    setGlobal('TEST_DOWNLOAD_TOOL_RETRY_MIN_SECONDS', undefined)
+    setGlobal('TEST_DOWNLOAD_TOOL_RETRY_MAX_SECONDS', undefined)
   })
 
   it('downloads a 35 byte file', async () => {
@@ -90,6 +100,7 @@ describe('@actions/tool-cache', function() {
 
   it('downloads a 35 byte file after a redirect', async () => {
     nock('http://example.com')
+      .persist()
       .get('/redirect-to')
       .reply(303, undefined, {
         location: 'http://example.com/bytes/35'
@@ -103,8 +114,75 @@ describe('@actions/tool-cache', function() {
     expect(fs.statSync(downPath).size).toBe(35)
   })
 
+  it('handles error from response message stream', async () => {
+    nock('http://example.com')
+      .persist()
+      .get('/error-from-response-message-stream')
+      .reply(200, {})
+
+    setResponseMessageFactory(() => {
+      const readStream = new stream.Readable()
+      /* eslint-disable @typescript-eslint/unbound-method */
+      readStream._read = () => {
+        readStream.destroy(new Error('uh oh'))
+      }
+      /* eslint-enable @typescript-eslint/unbound-method */
+      return readStream
+    })
+
+    let error = new Error('unexpected')
+    try {
+      await tc.downloadTool(
+        'http://example.com/error-from-response-message-stream'
+      )
+    } catch (err) {
+      error = err
+    }
+
+    expect(error).not.toBeUndefined()
+    expect(error.message).toBe('uh oh')
+  })
+
+  it('retries error from response message stream', async () => {
+    nock('http://example.com')
+      .persist()
+      .get('/retries-error-from-response-message-stream')
+      .reply(200, {})
+
+    /* eslint-disable @typescript-eslint/unbound-method */
+    let attempt = 1
+    setResponseMessageFactory(() => {
+      const readStream = new stream.Readable()
+      let failsafe = 0
+      readStream._read = () => {
+        // First attempt fails
+        if (attempt++ === 1) {
+          readStream.destroy(new Error('uh oh'))
+          return
+        }
+
+        // Write some data
+        if (failsafe++ === 0) {
+          readStream.push(''.padEnd(35))
+          readStream.push(null) // no more data
+        }
+      }
+
+      return readStream
+    })
+    /* eslint-enable @typescript-eslint/unbound-method */
+
+    const downPath = await tc.downloadTool(
+      'http://example.com/retries-error-from-response-message-stream'
+    )
+
+    expect(fs.existsSync(downPath)).toBeTruthy()
+    expect(fs.statSync(downPath).size).toBe(35)
+  })
+
   it('has status code in exception dictionary for HTTP error code responses', async () => {
     nock('http://example.com')
+      .persist()
       .get('/bytes/bad')
       .reply(400, {
         username: 'bad',
@@ -124,6 +202,7 @@ describe('@actions/tool-cache', function() {
 
   it('works with redirect code 302', async function() {
     nock('http://example.com')
+      .persist()
       .get('/redirect-to')
       .reply(302, undefined, {
         location: 'http://example.com/bytes/35'
@@ -256,7 +335,7 @@ describe('@actions/tool-cache', function() {
             .readFileSync(path.join(tempDir, 'mock7zr-args.txt'))
             .toString()
             .trim()
-        ).toBe(`x -bb1 -bd -sccUTF-8 ${_7zFile}`)
+        ).toBe(`x -bb0 -bd -sccUTF-8 ${_7zFile}`)
         expect(fs.existsSync(path.join(extPath, 'file.txt'))).toBeTruthy()
         expect(
           fs.existsSync(path.join(extPath, 'file-with-ç-character.txt'))
@@ -267,6 +346,110 @@ describe('@actions/tool-cache', function() {
       } finally {
         await io.rmRF(tempDir)
       }
+    })
+  } else if (IS_MAC) {
+    it('extract .xar', async () => {
+      const tempDir = path.join(tempPath, 'test-install.xar')
+      const sourcePath = path.join(__dirname, 'data', 'archive-content')
+      const targetPath = path.join(tempDir, 'test.xar')
+      await io.mkdirP(tempDir)
+
+      // Create test archive
+      const xarPath = await io.which('xar', true)
+      await exec.exec(`${xarPath}`, ['-cf', targetPath, '.'], {
+        cwd: sourcePath
+      })
+
+      // extract/cache
+      const extPath: string = await tc.extractXar(targetPath, undefined, '-x')
+      await tc.cacheDir(extPath, 'my-xar-contents', '1.1.0')
+      const toolPath: string = tc.find('my-xar-contents', '1.1.0')
+
+      expect(fs.existsSync(toolPath)).toBeTruthy()
+      expect(fs.existsSync(`${toolPath}.complete`)).toBeTruthy()
+      expect(fs.existsSync(path.join(toolPath, 'file.txt'))).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'file-with-ç-character.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'folder', 'nested-file.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.readFileSync(
+          path.join(toolPath, 'folder', 'nested-file.txt'),
+          'utf8'
+        )
+      ).toBe('folder/nested-file.txt contents')
+    })
+
+    it('extract .xar to a directory that does not exist', async () => {
+      const tempDir = path.join(tempPath, 'test-install.xar')
+      const sourcePath = path.join(__dirname, 'data', 'archive-content')
+      const targetPath = path.join(tempDir, 'test.xar')
+      await io.mkdirP(tempDir)
+
+      const destDir = path.join(tempDir, 'not-exist')
+
+      // Create test archive
+      const xarPath = await io.which('xar', true)
+      await exec.exec(`${xarPath}`, ['-cf', targetPath, '.'], {
+        cwd: sourcePath
+      })
+
+      // extract/cache
+      const extPath: string = await tc.extractXar(targetPath, destDir, '-x')
+      await tc.cacheDir(extPath, 'my-xar-contents', '1.1.0')
+      const toolPath: string = tc.find('my-xar-contents', '1.1.0')
+
+      expect(fs.existsSync(toolPath)).toBeTruthy()
+      expect(fs.existsSync(`${toolPath}.complete`)).toBeTruthy()
+      expect(fs.existsSync(path.join(toolPath, 'file.txt'))).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'file-with-ç-character.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'folder', 'nested-file.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.readFileSync(
+          path.join(toolPath, 'folder', 'nested-file.txt'),
+          'utf8'
+        )
+      ).toBe('folder/nested-file.txt contents')
+    })
+
+    it('extract .xar without flags', async () => {
+      const tempDir = path.join(tempPath, 'test-install.xar')
+      const sourcePath = path.join(__dirname, 'data', 'archive-content')
+      const targetPath = path.join(tempDir, 'test.xar')
+      await io.mkdirP(tempDir)
+
+      // Create test archive
+      const xarPath = await io.which('xar', true)
+      await exec.exec(`${xarPath}`, ['-cf', targetPath, '.'], {
+        cwd: sourcePath
+      })
+
+      // extract/cache
+      const extPath: string = await tc.extractXar(targetPath, undefined)
+      await tc.cacheDir(extPath, 'my-xar-contents', '1.1.0')
+      const toolPath: string = tc.find('my-xar-contents', '1.1.0')
+
+      expect(fs.existsSync(toolPath)).toBeTruthy()
+      expect(fs.existsSync(`${toolPath}.complete`)).toBeTruthy()
+      expect(fs.existsSync(path.join(toolPath, 'file.txt'))).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'file-with-ç-character.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.existsSync(path.join(toolPath, 'folder', 'nested-file.txt'))
+      ).toBeTruthy()
+      expect(
+        fs.readFileSync(
+          path.join(toolPath, 'folder', 'nested-file.txt'),
+          'utf8'
+        )
+      ).toBe('folder/nested-file.txt contents')
     })
   }
 
@@ -370,7 +553,9 @@ describe('@actions/tool-cache', function() {
       if (IS_WINDOWS) {
         const escapedStagingPath = stagingDir.replace(/'/g, "''") // double-up single quotes
         const escapedZipFile = zipFile.replace(/'/g, "''")
-        const powershellPath = await io.which('powershell', true)
+        const powershellPath =
+          (await io.which('pwsh', false)) ||
+          (await io.which('powershell', true))
         const args = [
           '-NoLogo',
           '-Sta',
@@ -383,7 +568,7 @@ describe('@actions/tool-cache', function() {
         ]
         await exec.exec(`"${powershellPath}"`, args)
       } else {
-        const zipPath: string = await io.which('zip')
+        const zipPath: string = await io.which('zip', true)
         await exec.exec(`"${zipPath}`, [zipFile, '-r', '.'], {cwd: stagingDir})
       }
 
@@ -421,7 +606,9 @@ describe('@actions/tool-cache', function() {
       if (IS_WINDOWS) {
         const escapedStagingPath = stagingDir.replace(/'/g, "''") // double-up single quotes
         const escapedZipFile = zipFile.replace(/'/g, "''")
-        const powershellPath = await io.which('powershell', true)
+        const powershellPath =
+          (await io.which('pwsh', false)) ||
+          (await io.which('powershell', true))
         const args = [
           '-NoLogo',
           '-Sta',
@@ -434,7 +621,7 @@ describe('@actions/tool-cache', function() {
         ]
         await exec.exec(`"${powershellPath}"`, args)
       } else {
-        const zipPath: string = await io.which('zip')
+        const zipPath: string = await io.which('zip', true)
         await exec.exec(zipPath, [zipFile, '-r', '.'], {cwd: stagingDir})
       }
 
@@ -478,7 +665,9 @@ describe('@actions/tool-cache', function() {
       if (IS_WINDOWS) {
         const escapedStagingPath = stagingDir.replace(/'/g, "''") // double-up single quotes
         const escapedZipFile = zipFile.replace(/'/g, "''")
-        const powershellPath = await io.which('powershell', true)
+        const powershellPath =
+          (await io.which('pwsh', false)) ||
+          (await io.which('powershell', true))
         const args = [
           '-NoLogo',
           '-Sta',
@@ -491,7 +680,7 @@ describe('@actions/tool-cache', function() {
         ]
         await exec.exec(`"${powershellPath}"`, args)
       } else {
-        const zipPath: string = await io.which('zip')
+        const zipPath: string = await io.which('zip', true)
         await exec.exec(zipPath, [zipFile, '-r', '.'], {cwd: stagingDir})
       }
 
@@ -541,4 +730,61 @@ describe('@actions/tool-cache', function() {
       expect(err.toString()).toContain('502')
     }
   })
+
+  it('retries 429s', async function() {
+    nock('http://example.com')
+      .get('/too-many-requests-429')
+      .times(2)
+      .reply(429, undefined)
+    nock('http://example.com')
+      .get('/too-many-requests-429')
+      .reply(500, undefined)
+
+    try {
+      const statusCodeUrl = 'http://example.com/too-many-requests-429'
+      await tc.downloadTool(statusCodeUrl)
+    } catch (err) {
+      expect(err.toString()).toContain('500')
+    }
+  })
+
+  it("doesn't retry 404", async function() {
+    nock('http://example.com')
+      .get('/not-found-404')
+      .reply(404, undefined)
+    nock('http://example.com')
+      .get('/not-found-404')
+      .reply(500, undefined)
+
+    try {
+      const statusCodeUrl = 'http://example.com/not-found-404'
+      await tc.downloadTool(statusCodeUrl)
+    } catch (err) {
+      expect(err.toString()).toContain('404')
+    }
+  })
 })
+
+/**
+ * Sets up a mock response body for downloadTool. This function works around a limitation with
+ * nock when the response stream emits an error.
+ */
+function setResponseMessageFactory(
+  factory: (() => stream.Readable) | undefined
+): void {
+  setGlobal('TEST_DOWNLOAD_TOOL_RESPONSE_MESSAGE_FACTORY', factory)
+}
+
+/**
+ * Sets a global variable
+ */
+function setGlobal<T>(key: string, value: T | undefined): void {
+  /* eslint-disable @typescript-eslint/no-explicit-any */
+  const g = global as any
+  /* eslint-enable @typescript-eslint/no-explicit-any */
+  if (value === undefined) {
+    delete g[key]
+  } else {
+    g[key] = value
+  }
+}